* Code cleanup before adding database + Added Docker MySQL image + Added JPA and MySQL Connector dependencies

+++ skm-client/src/main/java/pl/edu/pjwstk/skmclient/model/CompartmentStatus.java
@@ -7,17 +7,17 @@ import java.util.List;
 
 public class CompartmentStatus {
 
-    private final int id;
+    private final Long id;
     private final List<String> peopleNames;
 
     @JsonCreator
-    public CompartmentStatus(@JsonProperty("id") int id,
+    public CompartmentStatus(@JsonProperty("id") Long id,
                              @JsonProperty("peopleNames") List<String> peopleNames) {
         this.id = id;
         this.peopleNames = peopleNames;
     }
 
-    public int getID() {
+    public Long getID() {
         return id;
     }
